requirements.txt
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
inquirer==3.1.3
pytz==2019.3
python-dateutil==2.8.2
isodate==0.6.1
